backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic[email]==2.6.4
